toparchik_bot/requirements.txt
aiogram>=3.7.0,<4.0.0
aiohttp>=3.9.0
google-genai>=0.8.0
google-generativeai>=0.8.0
yt-dlp>=2025.1.15
curl-cffi>=0.7.0
python-dotenv>=1.0.0
python-pptx>=0.6.23
python-docx>=1.1.0
pdf2docx>=0.5.8
python-magic>=0.4.27
google-api-python-client>=2.120.0
